src/install: implement execution of pre/post-install handlers for each bundle
- Execute the new pre-install handler after the config-defined pre-install handler
- Execute the new post-install handler after the config-defined post-install handler
- Both handlers are executed in do_install_bundle()

Currently, RAUC supports pre/post-install handlers only via system.conf
defined in the old system. However, with this implementation, the old
system must predefine any actions that should be executed in the future.
With this change, actions can be defined using pre/post-install handlers
specified in the bundle itself, allowing actions to be executed before/after
all slots are updated. This enables handlers to be defined per bundle, so
the old system no longer needs to predefine future actions in its handlers.
This topic was also mentioned in issue #275 and #342.

Signed-off-by: Yosuke Tanimoto <Yosuke.A.Tanimoto@sony.com>